+++ unittest/lib/main.cpp
@@ -1316,6 +1316,7 @@ runFunctionalTests()
 
         testUtil();
 
+        testModuleInitEx();
         testScsTable();
 
         testScsTools();